+++ changelog.rst
@@ -3,7 +3,9 @@
 Change Log
 ##########
 
-- In development.
+- Version 0.4 (2023-01-06)
+  - Rename ``global-spell-fu-mode`` to ``spell-fu-global-mode``,
+    ``global-spell-fu-ignore-buffer`` to ``spell-fu-global-ignore-buffer``.
   - Add ``spell-fu-debug`` variable, to support debugging why dictionaries are not loading.
   - Add ``spell-fu-reset`` to re-generate cache.
   - Fix #31, failure to detect updated personal dictionary when it's a symlink.